Recoil 상태 관리에서 동승자 데이터 가져오기 방식을 usePassengers 훅으로 변경 및 실시간 구독 추가
